Modificaciones y agregaciones de componentes y rutas

+++ SCYGV/src/app/Components/usuarios/pag-usu/pag-usu.component.ts
@@ -1,26 +1,33 @@
 import { Component } from '@angular/core';
 import { FormGroup, FormBuilder, Validators, FormControl } from '@angular/forms';
+import { Rol } from 'src/app/modelo/rol';
 import { Usuario} from 'src/app/modelo/usuario';
+import { RolService } from 'src/app/services/rol.service';
 import { usuarioService } from 'src/app/services/usuario.service';
-
+import swal from'sweetalert2';
 @Component({
   selector: 'app-pag-usu',
   templateUrl: './pag-usu.component.html',
   styleUrls: ['./pag-usu.component.css']
 })
 export class PagUsuComponent {
- 
+
+  
+  LRoles !: Rol[];
   //Se utiliza para mostrar el usuario
   LUsuarios !: Usuario[];
   //Form
   FormUsu: FormGroup
-  
   //Iniciar el boton desactivado
-  isClicked: boolean = true;
+  
+  id:number = 0
 
-  constructor(private UService:usuarioService, private fb:FormBuilder){
+  mostrarC: boolean = true;
+  mostrarA: boolean = false;
+  constructor(private UService:usuarioService, private fb:FormBuilder, private rService : RolService){
     //Formulario para trabajar con los Usuarios
     this.FormUsu = fb.group({
+      id : new FormControl(),
       nombre : new FormControl('',[Validators.required]),
       apellidos : new FormControl('',[Validators.required]), 
       contrasena : new FormControl('',[Validators.required]),
@@ -34,18 +41,27 @@ export class PagUsuComponent {
       especialidad : new FormControl('',[Validators.required]),
       telefono : new FormControl('',[Validators.required])
     })
-
+  
     
   }
-  
   //Iniciar la muestra de Usuarios
   ngOnInit():void{
     this.getUsuario()
+    this.getRol()
   }
+
+   //Mostrar Rol
+getRol():any{
+  this.rService.getRol().subscribe(res =>{
+    this.LRoles = <any>res
+    console.log(res)
+  })
+}
 
   
   //Mostrar Usuarios
   getUsuario():any{
+    this.mostrarC = true;
     this.UService.getUsuario().subscribe(res =>{
       this.LUsuarios = <any>res
       console.log(res)})
@@ -70,16 +86,37 @@ export class PagUsuComponent {
       usuario.telefono = this.FormUsu.get('telefono')?.value
       this.UService.createUsuario(usuario).subscribe(res =>
         this.getUsuario(),
-        this.limpiar()
-      )}
+        this.limpiar())
+        swal.fire({
+        icon: 'success',
+        title: 'Se creo correctamente el Usuario',
+        showConfirmButton: false,
+        timer: 1500
+      })}
+     else{
+        swal.fire({
+          icon: 'error',
+          title: 'Oops...',
+          text: 'Faltan datos por rellenar...',})
+      }
   }
 
   //Eliminar Usuario
   deleteUsuario(idUsu : any){
-    this.UService.deleteUsuario(idUsu).subscribe(res =>{
+    swal.fire({
+        text: "¿Está seguro que desea eliminar?",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: "Sí, eliminar",
+        cancelButtonText: "Cancelar",
+    })
+    .then(resultado => {
+      if (resultado.value) {
+      this.UService.deleteUsuario(idUsu).subscribe(res =>{
       console.log(res)
       this.getUsuario()
     },error => console.log(error))
+  }})
   }
 
   //Actualizar Usuario
@@ -87,6 +124,7 @@ export class PagUsuComponent {
     if(this.FormUsu.valid)
     {
       let usuario = new Usuario()
+      usuario.id = this.FormUsu.get('id')?.value
       usuario.nombres = this.FormUsu.get('nombre')?.value
       usuario.apellidos = this.FormUsu.get('apellidos')?.value
       usuario.contrasena = this.FormUsu.get('contrasena')?.value
@@ -99,15 +137,18 @@ export class PagUsuComponent {
       usuario.f_ingreso = this.FormUsu.get('f_ingreso')?.value
       usuario.especialidad = this.FormUsu.get('especialidad')?.value
       usuario.telefono = this.FormUsu.get('telefono')?.value
+      idUsu = usuario.id;
       this.UService.updateUsuario(idUsu,usuario).subscribe(res =>
         this.getUsuario(),
-        this.limpiar(), 
-        this.Desactivar()
+        this.limpiar(),
       )}
     }
     
     //Funcion para editar el usuario
     editar(usuario:Usuario){
+      this.mostrarC = false;
+      this.mostrarA = true;
+      this.FormUsu.get('id')?.setValue(usuario.id)
       this.FormUsu.get('nombre')?.setValue(usuario.nombres)
       this.FormUsu.get('apellidos')?.setValue(usuario.apellidos)
       this.FormUsu.get('contrasena')?.setValue(usuario.contrasena)
@@ -120,7 +161,6 @@ export class PagUsuComponent {
       this.FormUsu.get('f_ingreso')?.setValue(usuario.f_ingreso)
       this.FormUsu.get('especialidad')?.setValue(usuario.especialidad)
       this.FormUsu.get('telefono')?.setValue(usuario.telefono)
-      this.Activar()
   }
   
   //Mostrar Usuario por Id
@@ -131,16 +171,6 @@ getUserId(idUsu: number):any{
   })
 }
   //Funcion para limpar el formulario
-  limpiar():any{this.FormUsu.reset()}
-  
-  //Funcion para activar boton Actualizar
-  Activar():any{this.isClicked = false}
-  
-  //Funcion para desactivar boton Actualizar
-  Desactivar():any{this.isClicked = true}
+  limpiar():any{this.mostrarC = true;this.mostrarA = false;this.FormUsu.reset()}
+ 
 }
-  
-  
-  
-    
-
